Maintenance
> Fixed various logical errors
> Code refactoring
> Added notification about not supporting a food

+++ app/src/main/java/com/iprism/elliot/data/local/entity/FoodIngredientCrossRef.kt
@@ -8,10 +8,10 @@ import org.jetbrains.annotations.NotNull
 data class FoodIngredientCrossRef(
     @ColumnInfo(name = "food_id")
     @NotNull
-    val foodId: Int,
+    val foodId: Long,
     @ColumnInfo(name = "ingredient_id")
     @NotNull
-    val ingredientId: Int,
+    val ingredientId: Long,
     @NotNull
     val lang: String
 )
