+++ app/src/main/java/com/iprism/elliot/domain/model/SubStatistic.kt
@@ -2,5 +2,5 @@ package com.iprism.elliot.domain.model
 
 data class SubStatistic(
     val statName: String,
-    val statEval: Double
+    val statEval: Float
 )
